@@ -1,8 +1,9 @@
 plugins {
-    id 'org.springframework.boot' version '2.6.3'
-    id 'io.spring.dependency-management' version '1.0.11.RELEASE'
+    id 'org.springframework.boot' version '3.2.5'
+    id 'io.spring.dependency-management' version '1.1.4'
     id 'java'
     id 'groovy'
+    id 'com.github.ben-manes.versions' version '0.42.0'
 }
 
 group = 'no.fintlabs'
@@ -27,15 +28,20 @@ repositories {
     mavenLocal()
 }
 
+java {
+    toolchain {
+        languageVersion = JavaLanguageVersion.of(21)
+    }
+}
+
 dependencies {
-    implementation 'com.google.guava:guava:28.2-jre'
-    implementation 'org.apache.commons:commons-lang3:3.9'
+    implementation 'com.google.guava:guava:33.0.0-jre'
+    implementation 'org.apache.commons:commons-lang3:3.14.0'
     implementation 'commons-beanutils:commons-beanutils:1.9.4'
-    implementation 'org.apache.jena:jena-arq:3.17.0'
-    implementation 'org.glassfish.jersey.core:jersey-client:2.29.1'
-    implementation 'org.glassfish.jersey.core:jersey-common:2.29.1'
-    implementation 'org.jooq:jool-java-8:0.9.14'
-    //implementation 'org.projectlombok:lombok:1.18.10'
+    implementation 'org.apache.jena:jena-arq:4.10.0'
+    implementation 'org.glassfish.jersey.core:jersey-client:2.35'
+    implementation 'org.glassfish.jersey.core:jersey-common:2.35'
+    implementation 'org.jooq:jool:0.9.15'
     compileOnly 'org.projectlombok:lombok'
     annotationProcessor 'org.projectlombok:lombok'
 
@@ -44,6 +50,10 @@ dependencies {
     implementation 'no.fint:fint-relation-model:1.1.6'
     implementation 'no.fint:fint-spring-secrets:0.2.2'
     implementation 'no.fint:fint-sse:2.0.0'
+    
+    // Java EE / Jakarta EE dependencies
+    implementation 'javax.ws.rs:javax.ws.rs-api:2.1.1'
+    implementation 'javax.annotation:javax.annotation-api:1.3.2'
 
     implementation "no.fint:fint-administrasjon-resource-model-java:${apiVersion}"
     implementation "no.fint:fint-utdanning-resource-model-java:${apiVersion}"
@@ -62,9 +72,14 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-data-mongodb'
 
-    testImplementation 'cglib:cglib-nodep:3.2.12'
-    testImplementation 'org.spockframework:spock-spring:1.1-groovy-2.4'
-    testImplementation 'org.spockframework:spock-core:1.1-groovy-2.4'
+    // Jersey dependencies for Java 21 compatibility
+    implementation 'org.glassfish.jersey.media:jersey-media-sse:2.35'
+    implementation 'org.glassfish.jersey.inject:jersey-hk2:2.35'
+
+    testImplementation 'cglib:cglib-nodep:3.3.0'
+    testImplementation 'org.spockframework:spock-spring:2.4-M1-groovy-4.0'
+    testImplementation 'org.spockframework:spock-core:2.4-M1-groovy-4.0'
+    testImplementation 'org.apache.groovy:groovy:4.0.17'
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
 }
 
@@ -80,23 +95,27 @@ jar {
     }
 }
 
-sourceCompatibility = 1.8
-
 import org.apache.tools.ant.filters.ReplaceTokens
 afterEvaluate {
     configure(processResources) {
         filesMatching('**/*.yml') {
-            filter(ReplaceTokens, tokens: [version: jar.version, title: jar.baseName])
+            filter(ReplaceTokens, tokens: [version: project.version, title: project.name])
         }
     }
 }
 
 bootRun {
+    javaLauncher = javaToolchains.launcherFor {
+        languageVersion = JavaLanguageVersion.of(21)
+    }
     systemProperties System.properties
 }
+
+// Fjernet den midlertidige deaktiveringen av tester
 
 test {
     testLogging {
         events 'passed', 'skipped', 'failed'
     }
+    useJUnitPlatform()
 }
